refactor: move module def files to mod-named files
