--- src/lib/api_app/src/api/routes/testbed/productizers.rs
@@ -1,2 +1,0 @@
-pub mod figure;
-pub mod job;
