+++ src/lib/api_app/src/api/routes/testbed/productizers/mod.rs
@@ -1,12 +1,14 @@
 use http::{HeaderMap, HeaderValue};
 
-use crate::api::utils::ParsedRequest;
-use crate::api::errors::APIRoutingError;
+use crate::api::{errors::APIRoutingError, utils::ParsedRequest};
+
+pub mod figure;
+pub mod job;
 
 /**
  * Parses the authorization headers fromn the input request
  */
-pub fn parse_testbed_request_headers(request: ParsedRequest) -> Result<HeaderMap, APIRoutingError> {
+fn parse_testbed_request_headers(request: ParsedRequest) -> Result<HeaderMap, APIRoutingError> {
     // Prep auth header forwarding
     let mut request_headers = HeaderMap::new();
     request_headers.insert("Content-Type", HeaderValue::from_static("application/json"));
@@ -14,14 +16,16 @@ pub fn parse_testbed_request_headers(request: ParsedRequest) -> Result<HeaderMap
         "authorization",
         request
             .headers
-            .get("authorization").ok_or(APIRoutingError::UnprocessableEntity)?
+            .get("authorization")
+            .ok_or(APIRoutingError::UnprocessableEntity)?
             .clone(),
     );
     request_headers.insert(
         "x-authorization-provider",
         request
             .headers
-            .get("x-authorization-provider").ok_or(APIRoutingError::UnprocessableEntity)?
+            .get("x-authorization-provider")
+            .ok_or(APIRoutingError::UnprocessableEntity)?
             .clone(),
     );
